# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TreeKB API
  version: "2.3.0"
  description: |
    TreeKB is an AI-driven, self-evolving knowledge graph system. It ingests documents,
    extracts knowledge points (KPs), organises them into a hierarchical topic tree, and
    provides intelligent question-answering with citations.

    **Key concepts**
    - **Dataset** -- an isolated knowledge base with its own SQLite database.
    - **Node** -- a topic node in the hierarchical tree.
    - **Chunk / KP** -- an atomic knowledge point stored under a node.
    - **Schema node** -- a curated node that guides ingestion mapping.
    - **Decision** -- a pending human-review item produced by the KP decision engine.

    Most endpoints require the `X-Dataset-ID` header to select the active dataset.
    Endpoints under `/datasets`, `/health`, `/metrics`, and `/settings` do not.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  parameters:
    DatasetIdHeader:
      name: X-Dataset-ID
      in: header
      required: true
      description: UUID of the target dataset. Resolved by middleware; falls back to default dataset when omitted.
      schema:
        type: string
        format: uuid
  schemas:
...
    SimpleError:
      type: object
      properties:
        error:
          type: string

    # ── Dataset ───────────────────────────────────────────────────────────────
...
    PromptEntry:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        category:
          type: string
          enum: [ingestion, query, routing]
        description:
          type: string
        variables:
          type: array
          items:
            type: string
        default_text:
          type: string
        current_text:
          type: string
        is_custom:
          type: boolean

    # ── Schema template ───────────────────────────────────────────────────────
...
paths:
...
  /prompts:
    get:
      operationId: listPrompts
      tags: [Prompts]
      summary: List all prompts
      description: Returns all prompt templates with their defaults and any per-dataset overrides.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: Prompt list
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompts:
                    type: array
                    items:
                      $ref: "#/components/schemas/PromptEntry"
  /prompts/reset:
    post:
      operationId: resetAllPrompts
      tags: [Prompts]
      summary: Reset all prompts to defaults
      description: Deletes all per-dataset prompt overrides, restoring every prompt to its built-in default.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
      responses:
        "200":
          description: All prompts reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /prompts/{key}:
    get:
      operationId: getPrompt
      tags: [Prompts]
      summary: Get a single prompt
      description: Returns the full details of a prompt template including default text, current text, and override status.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: key
          in: path
          required: true
          description: Prompt key (e.g. "kpExtraction_en", "answerGeneration_en")
          schema:
            type: string
      responses:
        "200":
          description: Prompt details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptEntry"
        "404":
          description: Unknown prompt key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

    put:
      operationId: savePromptOverride
      tags: [Prompts]
      summary: Save a prompt override
      description: Saves a custom prompt override for the active dataset. The override takes effect immediately for all subsequent LLM calls using this prompt key.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: The custom prompt text. Use {{variable}} placeholders for dynamic values.
      responses:
        "200":
          description: Override saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  key:
                    type: string
                  is_custom:
                    type: boolean
        "400":
          description: Missing or empty text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "404":
          description: Unknown prompt key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

    delete:
      operationId: resetPrompt
      tags: [Prompts]
      summary: Reset a prompt to default
      description: Deletes the per-dataset override for a single prompt, restoring its built-in default.
      parameters:
        - $ref: "#/components/parameters/DatasetIdHeader"
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Override removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  key:
                    type: string
                  is_custom:
                    type: boolean
                    example: false
        "404":
          description: Unknown prompt key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"

  # ── Manage (chatbot) ─────────────────────────────────────────────────────
...